extract Session object to a standalone class for reuse in retrofit framework
