+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
 
 subprojects {
     group 'org.ccci.gto.android'
-    version '0.8.1-SNAPSHOT'
+    version '0.9.0-SNAPSHOT'
 
     repositories {
         mavenLocal()
