+++ gto-support-api-base/build.gradle
@@ -1,0 +1,5 @@
+apply plugin: 'com.android.library'
+
+dependencies {
+    compile project(':gto-support-core')
+}
