+++ gto-support-api-base/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="org.ccci.gto.android.common.api.base"/>
